scripts/requirements.txt
httpx>=0.27.0
google-genai>=1.0.0
anthropic>=0.40.0
Pillow>=10.3.0
python-dotenv>=1.0.0
